Add information on running pipeline
Add dependencies and other information that would be necessary to run the pipeline from scratch.
Reformat the hard-coded directories to make them a bit more intuitive.
Delete some unused scripts.
Restore the "No Cas1/2" script.
