+++ src/nontn7/rules-nocas1-2.py
@@ -1,5 +1,4 @@
-""" Requires that systems have cas1 and cas2. """
-
+""" Excludes systems that have Cas1 or Cas2. """
 
 import sys
 
@@ -7,6 +6,5 @@ from operon_analyzer import analyze, rules
 
 from tools.filters import fs
 
-rs = rules.RuleSet().require('cas1').require('cas2')
-
+rs = rules.RuleSet().exclude('cas1').exclude('cas2')
 analyze.evaluate_rules_and_reserialize(sys.stdin, rs, fs)
